calculator divide 수정

+++ src/main/java/study/Calculator.java
@@ -14,8 +14,8 @@ public class Calculator {
 
         return a*b;
     }
-    public static int divide(int a,int b){
+    public static double divide(int a,int b){
 
-        return a/b;
+        return (double)a/b;
     }
 }
